server/requirements.txt
alembic==1.13.2
aniso8601==9.0.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
black==24.4.2
blinker==1.8.2
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.15.4
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
fsspec==2024.6.0
greenlet==3.0.3
huggingface-hub==0.23.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.3
mypy-extensions==1.0.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pgvector==0.2.5
platformdirs==4.2.2
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.31
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2